+++ Water_Features/Systems/FindWaterSourcesSystem.cs
@@ -71,21 +71,14 @@ namespace Water_Features.Systems
         /// <inheritdoc/>
         protected override void OnUpdate()
         {
-            bool seasonalStreamsEnabled = WaterFeaturesMod.Instance.Settings.EnableSeasonalStreams;
-            bool wavesAndTidesEnabled = WaterFeaturesMod.Instance.Settings.EnableWavesAndTides;
-            if (m_ToolSystem.actionMode.IsEditor())
+            if (!m_ToolSystem.actionMode.IsGameOrEditor())
             {
-                if (!WaterFeaturesMod.Instance.Settings.SeasonalStreamsAffectEditorSimulation)
-                {
-                    seasonalStreamsEnabled = false;
-                }
-
-                if (!WaterFeaturesMod.Instance.Settings.WavesAndTidesAffectEditorSimulation)
-                {
-                    wavesAndTidesEnabled = false;
-                }
+                Enabled = false;
+                return;
             }
 
+            bool seasonalStreamsEnabled = WaterFeaturesMod.Instance.Settings.EnableSeasonalStreams;
+            bool wavesAndTidesEnabled = WaterFeaturesMod.Instance.Settings.EnableWavesAndTides;
 
             m_Log.Debug($"{nameof(FindWaterSourcesSystem)}.{nameof(OnUpdate)} WaterFeaturesMod.Instance.Settings.EnableSeasonalStreams = {WaterFeaturesMod.Instance.Settings.EnableSeasonalStreams} &  WaterFeaturesMod.Instance.Settings.EnableWavesAndTides = {WaterFeaturesMod.Instance.Settings.EnableWavesAndTides}");
             FindWaterSourcesJob findWaterSourcesJob = new ()