+++ Water_Features/ChangeLog.md
@@ -1,5 +1,6 @@
 ﻿## Patch V1.1.11
 * Added a setting to attempt to force water simulation in game to be at least 1x.
 * Seasonal streams is by default off.
-* Seasonal Streams and Waves and Tides settings are only available in game or in the editor. Whether those options were last enabled or not in a save file determines whether the settings are enabled or disabled after loading. Valid water sources are required or it will default to disabled.
+* Seasonal Streams and Waves and Tides settings are only available in game or in the editor. All aspects of those settings are saved into save games now. Valid water sources are required or it will default to disabled.
 * Reorganized settings including making Water Tool tab as initial tab.
+* Additional settings to make waves and tides / seasonal streams work in the editor have been removed.